Updated RightClimber to have the configs of the CANSparkMax motor controller
The Right Climber should now have similar constructors and configs to the left Climber. The methods have not been changed yet but I will do that in the future. For now the constructors/configs between both should be the same and I will update the methods/commands in both to reflect it.